+++ example_runner.py
@@ -1,29 +1,32 @@
 
-print('Barrier examples')
-import example.barrier
-import example.barrier_alt
-import example.barrier_greebled
-import example.barrier_magnet_template
+print('Barrier Examples')
+import example.barrier.barrier
+import example.barrier.barrier_alt
+import example.barrier.barrier_greebled
+import example.barrier.barrier_magnet_template
+
+print('Bridge Examples')
+import example.bridge.straight
 
 print("Chip Tower Examples")
-import example.chip_can
-import example.combine
-import example.platformExample
-import example.ring
-import example.ring_conduit
-import example.guard_rail
-import example.guard_rail_platform
-import example.chip_tower_set
-import example.chip_tower_set_2
-import example.chiptower_readme_example
+import example.chip.chip_can
+import example.chip.combine
+import example.chip.platformExample
+import example.chip.ring
+import example.chip.ring_conduit
+import example.chip.guard_rail
+import example.chip.guard_rail_platform
+import example.chip.chip_tower_set
+import example.chip.chip_tower_set_2
+import example.chip.chiptower_readme_example
 
 print('Can Tower Examples')
-import example.can_platform
-import example.can_rail
-import example.can_tower
-import example.can_tower_set
+import example.can.can_platform
+import example.can.can_rail
+import example.can.can_tower
+import example.can.can_tower_set
 
-print('Walkway examples')
-import example.walkway
-import example.walkway_irregular
-import example.walkway_tile
+print('Walkway Examples')
+import example.walkway.walkway
+import example.walkway.walkway_irregular
+import example.walkway.walkway_tile