+++ src/cqindustry/can/__init__.py
@@ -1,0 +1,3 @@
+from .CanPlatform import CanPlatform
+from .CanRail import CanRail
+from .CanTower import CanTower